Pass the file path to the sidekiq worker instead of a file object
It is best practice when using Sidekiq to pass identifiers into
workers rather than objects. Making that change here prepares us to
make asynchronous calls to AssetManagerWorker.

The identifier for the uploaded file is the file's path. In the case
of the CarrierWave uploader the true path of the uploaded file (in
cases where a file has multiple versions) is the path of the
underlying File object (accessed using CarrierWave::SanitizedFile#to_file).