Make requests to Asset Manager API asynchronously
In the previous commits we changed the worker method signature to
accept a file path rather than a file object, and allowed the worker
and uploader to run on different machines by introducing a shared
`asset-manager-tmp` directory for temporary file storage.

The AssetManagerWorker can now be called asynchronously (with
`perform_async`).